refactor admin ip ban tasks
