+++ handlers/admin/add_ip_ban_task.go
@@ -1,0 +1,83 @@
+package admin
+
+import (
+	"database/sql"
+	"fmt"
+	"net/http"
+	"strings"
+	"time"
+
+	"github.com/arran4/goa4web/core/common"
+	"github.com/arran4/goa4web/core/consts"
+	"github.com/arran4/goa4web/handlers"
+	"github.com/arran4/goa4web/internal/db"
+	notif "github.com/arran4/goa4web/internal/notifications"
+	"github.com/arran4/goa4web/internal/tasks"
+)
+
+// AddIPBanTask blocks a network from accessing the site.
+type AddIPBanTask struct{ tasks.TaskString }
+
+var addIPBanTask = &AddIPBanTask{TaskString: TaskAdd}
+
+var _ tasks.Task = (*AddIPBanTask)(nil)
+var _ tasks.AuditableTask = (*AddIPBanTask)(nil)
+var _ notif.AdminEmailTemplateProvider = (*AddIPBanTask)(nil)
+
+func (AddIPBanTask) Action(w http.ResponseWriter, r *http.Request) any {
+	queries := r.Context().Value(consts.KeyCoreData).(*common.CoreData).Queries()
+	cd := r.Context().Value(consts.KeyCoreData).(*common.CoreData)
+
+	ipNet := strings.TrimSpace(r.PostFormValue("ip"))
+	ipNet = NormalizeIPNet(ipNet)
+	reason := strings.TrimSpace(r.PostFormValue("reason"))
+	expiresStr := strings.TrimSpace(r.PostFormValue("expires"))
+	var expires sql.NullTime
+	if expiresStr != "" {
+		if t, err := time.Parse("2006-01-02", expiresStr); err == nil {
+			expires = sql.NullTime{Time: t, Valid: true}
+		}
+	}
+	if ipNet != "" {
+		if err := queries.InsertBannedIp(r.Context(), db.InsertBannedIpParams{
+			IpNet:     ipNet,
+			Reason:    sql.NullString{String: reason, Valid: reason != ""},
+			ExpiresAt: expires,
+		}); err != nil {
+			return fmt.Errorf("insert banned ip fail %w", handlers.ErrRedirectOnSamePageHandler(err))
+		}
+	}
+	if evt := cd.Event(); evt != nil {
+		if evt.Data == nil {
+			evt.Data = map[string]any{}
+		}
+		evt.Data["IP"] = ipNet
+		if reason != "" {
+			evt.Data["Reason"] = reason
+		}
+		if u, _ := cd.CurrentUser(); u != nil {
+			evt.Data["Moderator"] = u.Username
+		}
+	}
+	return nil
+}
+
+func (AddIPBanTask) AdminEmailTemplate() *notif.EmailTemplates {
+	return notif.NewEmailTemplates("adminAddIPBanEmail")
+}
+
+func (AddIPBanTask) AdminInternalNotificationTemplate() *string {
+	v := notif.NotificationTemplateFilenameGenerator("adminAddIPBanEmail")
+	return &v
+}
+
+// AuditRecord summarises the addition of an IP ban.
+func (AddIPBanTask) AuditRecord(data map[string]any) string {
+	ip, _ := data["IP"].(string)
+	mod, _ := data["Moderator"].(string)
+	reason, _ := data["Reason"].(string)
+	if reason != "" {
+		return fmt.Sprintf("%s banned %s (%s)", mod, ip, reason)
+	}
+	return fmt.Sprintf("%s banned %s", mod, ip)
+}
